+++ ClassOverride/Updatable.js
@@ -1,0 +1,19 @@
+class Updatable{
+
+    constructor(){
+        
+    }
+
+    LoadOrig(){
+        
+    }
+    
+    Update(){
+        
+    }
+
+    Unload(){
+    }
+}
+
+export {Updatable};
