+++ ClassOverride/WithParent.js
@@ -1,0 +1,10 @@
+import { Updatable } from "./Updatable";
+
+class WithParent extends Updatable{
+    constructor(parent){
+        super();
+        this.parent = parent;
+    }
+}
+
+export {WithParent};
